requirements.txt
PyQt6==6.2.0
pytz==2021.3
tzlocal==4.0
pywin32==302
jstyleson==0.0.2
psutil==5.8.0
humanize==3.12.0
cssutils==2.3.0
PyYAML==6.0
Cerberus==1.3.4